+++ packages/headless/src/controllers/core/search-box/headless-core-search-box.ts
@@ -40,7 +40,10 @@ import {
   registerQuerySuggest,
   selectQuerySuggestion,
 } from '../../../features/query-suggest/query-suggest-actions';
-import {SearchAction} from '../../../features/analytics/analytics-utils';
+import {
+  InsightAction,
+  SearchAction,
+} from '../../../features/analytics/analytics-utils';
 import {prepareForSearchWithQuery} from '../../../features/search/search-actions';
 import {logQuerySuggestionClick} from '../../../features/query-suggest/query-suggest-analytics-actions';
 import {logSearchboxSubmit} from '../../../features/query/query-analytics-actions';
@@ -140,8 +143,10 @@ export interface SearchBox extends Controller {
 
   /**
    * Deselects all facets and triggers a search query.
+   *
+   * @param analytics -  The analytics action to log after submitting a query.
    */
-  submit(): void;
+  submit(analytics?: SearchAction): void;
 
   /**
    * The state of the `SearchBox` controller.
@@ -231,8 +236,8 @@ export function buildCoreSearchBox(
       );
     },
 
-    submit() {
-      performSearch(logSearchboxSubmit());
+    submit(analytics: SearchAction | InsightAction = logSearchboxSubmit()) {
+      performSearch(analytics);
       dispatch(clearQuerySuggest({id}));
     },