+++ packages/atomic-react/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.24.6](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.24.5...@coveo/atomic-react@1.24.6) (2022-09-19)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.24.5](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.24.4...@coveo/atomic-react@1.24.5) (2022-09-16)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.24.4](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.24.3...@coveo/atomic-react@1.24.4) (2022-09-16)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.24.3](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.24.2...@coveo/atomic-react@1.24.3) (2022-09-16)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.24.2](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.24.1...@coveo/atomic-react@1.24.2) (2022-09-15)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
 ## [1.24.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.24.0...@coveo/atomic-react@1.24.1) (2022-09-15)
 
 **Note:** Version bump only for package @coveo/atomic-react