+++ packages/atomic/CHANGELOG.md
@@ -3,6 +3,65 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.112.2](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.112.1...@coveo/atomic@1.112.2) (2022-09-19)
+
+
+### Bug Fixes
+
+* **atomic:** fix default template for atomic-result-text on language ([#2447](https://github.com/coveo/ui-kit/issues/2447)) ([6b24146](https://github.com/coveo/ui-kit/commit/6b2414644a871e0e39bf0791b3a7f51908241344))
+
+
+
+
+
+## [1.112.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.112.0...@coveo/atomic@1.112.1) (2022-09-16)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+# [1.112.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.111.0...@coveo/atomic@1.112.0) (2022-09-16)
+
+
+### Bug Fixes
+
+* **atomic:** render result list content when it's ready ([#2438](https://github.com/coveo/ui-kit/issues/2438)) ([0f62dcd](https://github.com/coveo/ui-kit/commit/0f62dcd4ea66b35033fc1199b59f216cbe718b02))
+
+
+### Features
+
+* **atomic:** added support for multiple notifications ([#2437](https://github.com/coveo/ui-kit/issues/2437)) ([a40dce2](https://github.com/coveo/ui-kit/commit/a40dce2cfa68e101e0e03d901673f192468c1993))
+
+
+
+
+
+# [1.111.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.110.1...@coveo/atomic@1.111.0) (2022-09-16)
+
+
+### Features
+
+* **atomic:** add error message when result-value-text is used with multi value field ([#2430](https://github.com/coveo/ui-kit/issues/2430)) ([4b9ea3e](https://github.com/coveo/ui-kit/commit/4b9ea3e26455767e836026b9bb87f73543fe561d))
+* **stencil:** make SPA with a single interface, switching search box type ([#2431](https://github.com/coveo/ui-kit/issues/2431)) ([cf5d01c](https://github.com/coveo/ui-kit/commit/cf5d01c7add8ed68858af0494251885ae971eb1a))
+
+
+
+
+
+## [1.110.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.110.0...@coveo/atomic@1.110.1) (2022-09-15)
+
+
+### Bug Fixes
+
+* **atomic:** added smart snippet inline link analytics ([#2417](https://github.com/coveo/ui-kit/issues/2417)) ([bb8cf73](https://github.com/coveo/ui-kit/commit/bb8cf73a35f02663a6abecad0dbeb598bfba1426))
+* **atomic:** fixed combobox ARIA attributes ([#2433](https://github.com/coveo/ui-kit/issues/2433)) ([76c0ed7](https://github.com/coveo/ui-kit/commit/76c0ed73e9bcff7634f8e276b592b89f0b777cac))
+
+
+
+
+
 # [1.110.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.109.0...@coveo/atomic@1.110.0) (2022-09-15)